Separate work
Moved the cleaning db logic to a new file.
notionSync file would rather check if a commit exist before adding it to db.
